renaming fetchutils.ts part 1
